requirements.txt
langchain-core
langchain-groq
python-dotenv
rich
psutil